requirements.txt
Django~=4.1.1
pytest~=7.1.3
django-ckeditor~=6.5.1
requests~=2.28.1
djangorestframework~=3.14.0